+++ home.nix
@@ -6,10 +6,18 @@ in {
   # Replace `with pkgs;` with `inherit (pkgs)`
   # https://nix.dev/anti-patterns/language#with-attrset-expression
   home.packages = builtins.attrValues {
+    # Necessary for non-NixOS systems which won't have the flakiest version of Nix
+    nixFlakes = (assert (builtins.compareVersions pkgs.nix.version "2.4") < 0; pkgs.nixFlakes);
+
     inherit (pkgs) peco ripgrep htop ranger;
 
     inherit (pkgs) _1password-gui qalculate-gtk pavucontrol;
   };
+
+  xdg.configFile."nix/nix.conf".text = (assert (builtins.compareVersions pkgs.nix.version "2.4") < 0; ''
+    experimental-features = nix-command flakes
+  '');
+
 
   # Allow fonts to be specified in `home.packages`
   fonts.fontconfig.enable = true;
@@ -99,6 +107,10 @@ in {
 
   programs.zsh = {
     enable = true;
+    # If this option is not disabled
+    # `home-manager` installs `nix-zsh-completions`
+    # which conflicts with `nix` in `home.packages`
+    enableCompletion = false;
 
     initExtraFirst = ''
       path=(